metrics_consumer/requirements.txt
kafka-python==1.4.7
psycopg2==2.8.4
SQLAlchemy==1.3.11
